refactor(controladores,intemerdiario, suporte) Refatoração para dry, evitando duplicação de códigos, códigos usados em suporte
